% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_appendix.R
\name{create_appendix}
\alias{create_appendix}
\title{Create an Appendix File for Your Academic Project}
\usage{
create_appendix(path)
}
\arguments{
\item{path}{a path in which to place the appendix file. Defaults to current
working directory in the absence of a user-specified argument. User-supplied
arguments here should be understood as subdirectories relative to current
working directory. Assume that the current working directory is something
like \verb{/home/steve/}, then specifying a path of "Documents" in this argument
would create the memo file in \verb{/home/steve/Documents}. Just be mindful about
your setup if you experiment with this argument.}
}
\value{
\code{create_appendix()} creates a \code{appendix.Rmd} file in the path
requested by user. Users can create an appendix at any time during the
submission timeline of the academic project, but I often---but not
always---create them at the revise-and-resubmit phase. That would explain
why it's not a default file.
}
\description{
\code{create_appendix()} creates an R Markdown for an appendix
to go with your academic project.
}
\examples{

\dontrun{
# Creates a `appendix.Rmd` file in working directory
create_appendix()

# Creates a `appendix.Rmd` in subdirectory of working directory

create_appendix("subdirectory")
}

}
