% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_memo.R
\name{create_memo}
\alias{create_memo}
\title{Create a Memo File for Your Academic Project}
\usage{
create_memo(path)
}
\arguments{
\item{path}{a path in which to place the memo file. Defaults to current
working directory in the absence of a user-specified argument. User-supplied
arguments here should be understood as subdirectories relative to current
working directory. Assume that the current working directory is something
like \verb{/home/steve/}, then specifying a path of "Documents" in this argument
would create the memo file in \verb{/home/steve/Documents}. Just be mindful about
your setup if you experiment with this argument.}
}
\value{
\code{create_memo()} creates a \code{memo.Rmd} file in the path
requested by user. Users can create a memo at any time during the submission
timeline of the academic project, but I think of these documents as
typically something a user writes when offered an opportunity to revise and
resubmit a manuscript. That would explain why the skeleton file is
structured the way it is.
}
\description{
\code{create_memo()} creates an R Markdown for a memo to go
with your academic project.
}
\examples{

\dontrun{
# Creates a `memo.Rmd` file in working directory
create_memo()

# Creates a `memo.Rmd` in subdirectory of working directory

create_memo("subdirectory")
}

}
