% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create an Academic Project/Paper Project (i.e. a Steve Project)}
\usage{
create_project(path, ...)
}
\arguments{
\item{path}{the directory name for the project (e.g. "my-project", "dissertation", or whatever)}

\item{...}{optional stuff, but you can ignore this}
}
\value{
\code{create_project()} creates a directory named with whatever is supplied in the \code{path} variable. This
directory will appear in the current working directory if it is executed in the command line. The contents of that
directory will include a skeleton project to get the user started. The user can (and must) ultimately make it their own.
}
\description{
\code{create_project()} provides a convenient function for creating
a new academic/paper project.
}
\examples{

\dontrun{
create_project("example")
create_project("dissertation")
}

}
