% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_appendix.R
\name{render_appendix}
\alias{render_appendix}
\title{Render Your R Markdown Appendix to Various Outputs}
\usage{
render_appendix(
  file = "appendix.Rmd",
  output_dir = "doc",
  outputs = c("pdf", "pdf-anon"),
  parameters,
  latex_engine = "xelatex",
  dev = "cairo_pdf"
)
}
\arguments{
\item{file}{the name of the R Markdown file containing that is your
appendix. Defaults to "appendix.Rmd".}

\item{output_dir}{the output directory to contain the formatted manuscript.
Defaults to "doc". Optimally, this is a subdirectory of the directory
containing the manuscript. A user who wants the formatted manuscript
to be in the same directory as the R Markdown file should specify
\code{output_dir = NULL} here.}

\item{outputs}{the various formatted manuscript types the user wants,
supplied as a character vector. Must be one or more of "pdf", "pdf-anon",
"word", and/or "html". No other formats are supported right now. Defaults
are "pdf" and "pdf-anon".}

\item{parameters}{optional parameters, specified as a character, based on
what's in your R Markdown file, passed as \code{params} in the \code{render()}
function in R Markdown. If no parameters are specified here, the
function defaults these parameters to
\verb{anonymous=TRUE, doublespacing=TRUE},
which is then wrapped in a list to be passed to the \code{params}
argument in \code{render()}. Do note this primarily concerns the anonymous
manuscript type. These are somewhat advanced-level arguments, so the user
should be careful what they do here and should have a firm idea what
they are doing here.}

\item{latex_engine}{the LaTeX engine the user may want to use.
Defaults to \code{"xelatex"}. You can overwrite this if you would like, but
why would you?}

\item{dev}{the graphics device for LaTeX PDFs. Defaults to \code{"cairo_pdf"}.
You can overwrite this, but why would you?}
}
\value{
\code{render_ms()} takes various arguments, most effectively built
into the function, and renders your R Markdown manuscript to various outputs
supported by \pkg{stevetemplates}.
}
\description{
\code{render_appendix()} takes various arguments, most effectively built
into the function, and renders your R Markdown manuscript to various outputs
supported by \pkg{stevetemplates}.
}
\examples{

\dontrun{
render_appendix()
}

}
