% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_memo.R
\name{render_memo}
\alias{render_memo}
\title{Render Your R Markdown Memo to Various Outputs}
\usage{
render_memo(
  file = "memo.Rmd",
  output_dir = "doc",
  outputs = c("pdf", "word"),
  latex_engine = "xelatex",
  dev = "cairo_pdf"
)
}
\arguments{
\item{file}{the name of the R Markdown file containing that is your
memo Defaults to "memo.Rmd".}

\item{output_dir}{the output directory to contain the formatted manuscript.
Defaults to "doc". Optimally, this is a subdirectory of the directory
containing the manuscript. A user who wants the formatted manuscript
to be in the same directory as the R Markdown file should specify
\code{output_dir = NULL} here.}

\item{outputs}{the various formatted manuscript types the user wants,
supplied as a character vector. Must be one or more of "pdf" and/or "word".
No other formats are supported right now.}

\item{latex_engine}{the LaTeX engine the user may want to use.
Defaults to \code{"xelatex"}. You can overwrite this if you would like, but
why would you?}

\item{dev}{the graphics device for LaTeX PDFs. Defaults to \code{"cairo_pdf"}.
You can overwrite this, but why would you?}
}
\value{
\code{render_memo()} takes various arguments, most effectively built
into the function, and renders your R Markdown memo to various outputs
supported by \pkg{stevetemplates}.
}
\description{
\code{render_memo()} takes various arguments, most effectively built
into the function, and renders your R Markdown memo to various outputs
supported by \pkg{stevetemplates}.
}
\examples{

\dontrun{
render_memo()
}

}
