% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pdfs.R
\name{render_pdfs}
\alias{render_pdfs}
\title{Render PDFs}
\usage{
render_pdfs(...)
}
\arguments{
\item{...}{optional stuff, but you can ignore this, for now}
}
\value{
dsafsdafsa
}
\description{
dsfasdfasfda
}
\examples{

\dontrun{
render_pdf()
}

}
